requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.1.0
sqlalchemy>=2.0.22
alembic>=1.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.2
openai>=1.2.0
python-multipart>=0.0.6
watchfiles>=0.21.0
tenacity>=8.2.3
structlog>=23.1.0
asyncio>=3.4.3
aiosqlite>=0.19.0
redis>=4.2.0
openai-agents>=0.0.9

# The modern way to use Redis with asyncio is directly via `redis.asyncio`:
# This is already correctly implemented in your cache.py file. The asyncio functionality was incorporated into the main redis package, eliminating the need for a separate redis-asyncio package. 